backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
boto3==1.35.94
ffmpeg-python==0.2.0
openai==1.59.6
supabase==2.11.1
python-jose[cryptography]==3.3.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0
cryptography==44.0.0
requests==2.32.3
